got a working rotation fix
rotation fix still has a few kinks, but it works and the game is playable now with some form of proper block rotations:

The fix:
It seems as though trying to rotate the block in every direction before giving up on the rotation was wrong. apparently you either rotate the direction you want, or the opposite way, or else no rotation, but I was trying to rotate the block a third time (180 degrees from original position before first rotation) and that wasn't supposed to be a feature. Also moving the block down to fix a rotation isn't good, so that was removed. Streamlined and refactored rotation code no to have 3 nested while-loops which was also causing issues.