Movement and row clear updates
Hold move right and left not working, it is glitchy when you spam left and right so I am commenting the current code out for now and am moving on from that for the time being with just single move taps being the movement option right now.

Row clear not working either, currently the raycast might be broken as no blocks are being picked up and added to the list of rows that need to be removed, debugging required!