ROW CLEAR FIX
Row clear previous issue was giving me issues where the proper blocks weren't getting cleared. The fix was that the raycast was hitting 2D box colliders in the parent of the blocks, instead of the children, so I created single square block prefabs and replaced all the children in all the block types with that prefab which had the collider, and removed the colliders from the parent, so the raycast now hit the children instead of the parent, and then I could use that raycast hit information to get the proper children to remove for row clears. :+1: