Camera shake and bug fixes
Added camera shake - made a static class camera shake with a camera shake component script, so camera shake is able to be called from anywhere and it finds the camera with a camera shake component and shakes it according to passed elements in the function.

fixed high score data saver object being loaded multiple times when going back to main menu because the object was manually placed in the level instead of instantiated with a check to see if it is already in the level.